Moved angular project files up
