+++ src/app/app.component.html
@@ -1,0 +1,14 @@
+<div class="container mt-4">
+  <ul class="nav nav-tabs">
+    <li class="nav-item">
+      <a class="nav-link" routerLink="/happiness" routerLinkActive="active">Happiness</a>
+    </li>
+    <li class="nav-item">
+      <a class="nav-link" routerLink="/research" routerLinkActive="active">Research</a>
+    </li>
+  </ul>
+
+  <div class="tab-content mt-2">
+    <router-outlet></router-outlet>
+  </div>
+</div>
