+++ src/app/interfaces/store-service.ts
