+++ src/app/models/city-info.ts
