+++ src/app/services/empire-info.service.ts
