+++ src/app/services/happiness-calculator.service.ts
