+++ src/app/services/housing-calculator.service.ts
