--- src/src/app/app.component.ts
@@ -1,16 +1,0 @@
-import {Component} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {RouterLink, RouterLinkActive, RouterOutlet} from '@angular/router';
-import {HappinessComponent} from "./components/happiness/happiness.component";
-import {ResearchComponent} from "./components/research/research.component";
-
-@Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [CommonModule, RouterOutlet, HappinessComponent, ResearchComponent, RouterLink, RouterLinkActive],
-  templateUrl: './app.component.html',
-  styleUrl: './app.component.css',
-})
-export class AppComponent {
-  title = 'icalcariam';
-}
